requirements.txt
#-e .
#pytest
aws-cdk.core==1.84.0
aws-cdk.aws-codecommit==1.84.0
aws-cdk.aws-codepipeline==1.84.0
aws-cdk.aws-codepipeline-actions==1.84.0
aws-cdk.pipelines==1.84.0
aws-cdk.aws-codegurureviewer==1.84.0
#aws-cdk.aws-s3==1.84.0
aws-cdk.aws-sqs==1.84.0
#enable bridgecrew for cdk synth checks
bridgecrew
#enable bandit for lambda checks
bandit